New upload and correction
MinAngleClassification is for trying to find the minimum number of angles for training the data to get a similar test accuracy for all of the methods used for the the whole training set. It is clear that a single angle is not nearly enough. Also, got rid of the seed for the random choice of images as we do not want to always pick the same images and even when we do, due to the nature of the models we use the accuracies are not the same